+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- [breaking] Upgraded to the latest React context API
 - Refactoring out deprecated react lifecycle methods
 - Upgrading to `@babel/*@7`